pyproject.toml
[tool.poetry]
name = "bb4-parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
pytest = "^8.2.2"
requests = "^2.32.3"
pytest-mock = "^3.14.0"
pytelegrambotapi = "^4.21.0"
coverage = "^7.5.4"
schedule = "^1.2.2"
setuptools = "^70.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
